parseMagstimResponse now with Double

+++ magstim.h
@@ -95,7 +95,7 @@ public:
     * @param[in] message                    [opt] TODO Doxygen
     * @param[in] error                      [opt] Chatches error code
     */
-    virtual void remoteControl(bool enable, std::map<QString, std::map<QString, int> > &message, int &error);
+    virtual void remoteControl(bool enable, std::map<QString, std::map<QString, double> > &message, int &error);
 
     //=========================================================================================================
     /**
@@ -109,7 +109,7 @@ public:
     * @param[in] message                    [opt] TODO Doxygen
     * @param[in] error                      [opt] Chatches error code
     */
-    virtual void arm(bool delay, std::map<QString, std::map<QString, int> > &message, int &error);
+    virtual void arm(bool delay, std::map<QString, std::map<QString, double> > &message, int &error);
 
     //=========================================================================================================
     /**
@@ -118,7 +118,7 @@ public:
     * @param[in] message                    [opt] TODO Doxygen
     * @param[in] error                      [opt] Chatches error code
     */
-    virtual void disarm(std::map<QString, std::map<QString, int> > &message, int &error);
+    virtual void disarm(std::map<QString, std::map<QString, double> > &message, int &error);
 
     //=========================================================================================================
     /**
@@ -127,7 +127,7 @@ public:
     * @param[in] error                      [opt] Chatches error code
     * @param[out] message                   TODO Doxygen
     */
-    virtual std::map<QString, std::map<QString, int> > getParameters(int &error);
+    virtual std::map<QString, std::map<QString, double> > getParameters(int &error);
 
     //=========================================================================================================
     /**
@@ -138,7 +138,7 @@ public:
     * @param[in] error                      [opt] Chatches error code
     * @param[out] message                   TODO Doxygen
     */
-    virtual std::map<QString, std::map<QString, int> > getTemperature(int &error);
+    virtual std::map<QString, std::map<QString, double> > getTemperature(int &error);
 
     //=========================================================================================================
     /**
@@ -175,7 +175,7 @@ public:
     * @param[in] commandByte                [opt] should not be changed by the user
     * @param[in] message                    [opt] TODO Doxygen
     */
-    virtual void setPower(int newPower, bool delay, int &error, QString commandByte, std::map<QString, std::map<QString, int> > &message);
+    virtual void setPower(int newPower, bool delay, int &error, QString commandByte, std::map<QString, std::map<QString, double> > &message);
 
     //=========================================================================================================
     /**
@@ -198,7 +198,7 @@ public:
     * @param[in] message                    [opt] TODO Doxygen
     * @param[in] error                      [opt] Chatches error code
     */
-    virtual void fire(std::map<QString, std::map<QString, int> > &message, int &error);
+    virtual void fire(std::map<QString, std::map<QString, double> > &message, int &error);
 
     //=========================================================================================================
     /**
@@ -246,7 +246,7 @@ protected:
     *
     * @return TODO Doxygen
     */
-    virtual std::map<QString, std::map<QString, int>> parseMagstimResponse(std::list<int> responseString, QString responseType);
+    virtual std::map<QString, std::map<QString, double>> parseMagstimResponse(std::list<int> responseString, QString responseType);
 
     //=========================================================================================================
     /**
@@ -275,7 +275,7 @@ protected:
     * @param[in] readBytes                  number of bytes in the response
     * @param[in] message                    TODO Doxygen
     */
-    virtual int processCommand(QString commandString, QString receiptType, int readBytes, std::map<QString, std::map<QString, int> > &message);
+    virtual int processCommand(QString commandString, QString receiptType, int readBytes, std::map<QString, std::map<QString, double> > &message);
 
     //=========================================================================================================
     /**
@@ -298,7 +298,7 @@ protected:
     * @param[in] version                    version of Magstim unit (X, Y, Z)
     * @param[in] message                    TODO Doxygen
     */
-    virtual int processCommand(QString commandString, QString receiptType, int readBytes, std::tuple<int, int, int> &version, std::map<QString, std::map<QString, int> > &message);
+    virtual int processCommand(QString commandString, QString receiptType, int readBytes, std::tuple<int, int, int> &version, std::map<QString, std::map<QString, double> > &message);
 
     //=========================================================================================================
     /**
@@ -312,7 +312,7 @@ protected:
 
     static int er;                                                    /**< Use as placeholder */
     static std::tuple<int, int, int> ver;                             /**< Use as placeholder */
-    static std::map<QString, std::map<QString, int>> mes;             /**< Use as placeholder */
+    static std::map<QString, std::map<QString, double>> mes;             /**< Use as placeholder */
 
     std::queue<std::tuple<QByteArray, QString, int>> m_sendQueue;     /**< Queue for writing to Magstim */
     std::queue<std::tuple<int, QByteArray>> m_receiveQueue;           /**< Queue for reading from Magstim */
